[Front] Deletion of a file -> deletion of the associated external reference + FileExportViewer in tsx pure function (#2417)
