+++ opencti-platform/opencti-front/src/schema/relay.schema.graphql
@@ -825,6 +825,7 @@ type File {
   metaData: FileMetadata
   uploadStatus: State!
   works: [Work]
+  externalReferenceId: String
 }
 
 type FileEdge {
@@ -1621,6 +1622,7 @@ enum ExternalReferencesFilter {
   usedBy
   source_name
   external_id
+  fileId
 }
 
 input ExternalReferencesFiltering {
